tideup the file structure of M4RI-based WBSM4
